pyproject.toml
[tool.poetry]
name = "volunteer-recognition-system"
version = "0.1.0"
description = ""
authors = ["Sasha <s.watanabe@ufl.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest-flask = "^1.3.0"
sqlalchemy = "^2.0.36"
flask-login = "^0.6.3"
flask-bcrypt = "^1.0.1"
flask-wtf = "^1.2.1"
flask-migrate = "^4.0.7"
selenium = "^4.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
